Apply file move feedback
